+++ src/main/java/com/github/games647/lagmonitor/LagMonitor.java
@@ -1,5 +1,6 @@
 package com.github.games647.lagmonitor;
 
+import com.github.games647.lagmonitor.commands.EnvironmentCommand;
 import com.github.games647.lagmonitor.commands.GraphCommand;
 import com.github.games647.lagmonitor.commands.MbeanCommand;
 import com.github.games647.lagmonitor.commands.MonitorCommand;
@@ -25,6 +26,7 @@ public class LagMonitor extends JavaPlugin {
         getCommand("tpshistory").setExecutor(new TpsHistoryCommand(this));
         getCommand("mbean").setExecutor(new MbeanCommand(this));
         getCommand("system").setExecutor(new SystemCommand(this));
+        getCommand("system").setExecutor(new EnvironmentCommand(this));
         getCommand("monitor").setExecutor(new MonitorCommand(this));
         getCommand("timing").setExecutor(new TimingCommand(this));
         getCommand("graph").setExecutor(new GraphCommand(this));