+++ src/main/java/com/github/games647/lagmonitor/commands/SystemCommand.java
@@ -3,11 +3,10 @@ package com.github.games647.lagmonitor.commands;
 import com.github.games647.lagmonitor.LagMonitor;
 
 import java.lang.management.ManagementFactory;
-import java.lang.management.OperatingSystemMXBean;
 import java.lang.management.RuntimeMXBean;
 import java.text.SimpleDateFormat;
 
-import org.apache.commons.lang.SystemUtils;
+import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
@@ -28,15 +27,11 @@ public class SystemCommand implements CommandExecutor {
 
         Runtime runtime = Runtime.getRuntime();
         sender.sendMessage(ChatColor.DARK_GREEN + "Uptime: " + uptimeFormat);
-        sender.sendMessage(ChatColor.DARK_GREEN + "Java: " + runtimeBean.getVmVersion());
         sender.sendMessage(ChatColor.DARK_GREEN + "Free RAM (MB): " + convertBytesToMegaBytes(runtime.freeMemory()));
-        sender.sendMessage(ChatColor.DARK_GREEN + "Cores: " + runtime.availableProcessors());
-        sender.sendMessage(ChatColor.DARK_GREEN + "Java: " + SystemUtils.JAVA_VERSION_FLOAT);
         sender.sendMessage(ChatColor.DARK_GREEN + "Threads: " + Thread.getAllStackTraces().size());
         sender.sendMessage(ChatColor.DARK_GREEN + "TPS: " + plugin.getTpsHistoryTask().getLastTicks());
 
-        OperatingSystemMXBean operationBean = ManagementFactory.getOperatingSystemMXBean();
-        sender.sendMessage(ChatColor.DARK_GREEN + "CPU Usage: " + operationBean.getSystemLoadAverage());
+        sender.sendMessage(ChatColor.DARK_GREEN + "Server version: " + Bukkit.getBukkitVersion());
         return true;
     }