+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
-# 0.1.1
+## 0.2
+
+* Added environment command
+* Added server version to the system command
+
+## 0.1.1
 
 * Added command permissions
+* Added online check for the ping command